Refactor project and update test suite
This commit includes a significant refactoring of the project. The solution file `ChadFairlie_ST10269509_PROG6221_POE.sln` has been updated to change the project names. Several files have been removed including `RecipeTests.cs`, `PROG6221_POE.Tests.csproj`, `AssemblyInfo.cs`, `packages.config`, `App.config`, `Ingredient.cs`, `Recipe.cs`, and `UnitConverter.cs`. The entire content of the `PROG6221_POE.csproj` and `Program.cs` files have also been removed.

A new `.csproj` file for the test project "ChadFairlie_PROG6221_POE.Tests.csproj" was added, along with a new `packages.config` file, a new `RecipeTests.cs` file, and a new `App.config` file. The `AssemblyInfo.cs` file was modified to add the assembly information related to the test project "PROG6221_POE.Tests".

New classes `Ingredient`, `Recipe`, and `UnitConverter` were added. The `.csproj` file for the project "ChadFairlie_PROG6221_POE.csproj" was modified to include the configuration for the project. The `AssemblyInfo.cs` file was added to provide general information about the assembly.
