+++ ChadFairlie_ST10269509_PROG6221_POE.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.10.34804.81
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PROG6221_POE", "PROG6221_POE\PROG6221_POE.csproj", "{0C529904-3DB1-407E-97C5-C9399238B76D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChadFairlie_PROG6221_POE", "ChadFairlie_PROG6221_POE\ChadFairlie_PROG6221_POE.csproj", "{0C529904-3DB1-407E-97C5-C9399238B76D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PROG6221_POE.Tests", "PROG6221_POE.Tests\PROG6221_POE.Tests.csproj", "{F8FCCE65-11EA-4821-99C5-F13B1C1E6A4C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChadFairlie_PROG6221_POE.Tests", "ChadFairlie_PROG6221_POE.Tests\ChadFairlie_PROG6221_POE.Tests.csproj", "{F8FCCE65-11EA-4821-99C5-F13B1C1E6A4C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
